Allow more than just .anon and .auth
Initial refactoring of SQL to allow more roles.

Conflicts:
	kernel-impl/src/main/java/org/sakaiproject/authz/impl/BaseAuthzGroup.java
	kernel-impl/src/main/java/org/sakaiproject/authz/impl/BaseAuthzGroupService.java
	kernel-impl/src/main/java/org/sakaiproject/authz/impl/DbAuthzGroupService.java
	kernel-impl/src/main/java/org/sakaiproject/authz/impl/DbAuthzGroupSql.java
	kernel-impl/src/main/java/org/sakaiproject/authz/impl/DbAuthzGroupSqlDefault.java
	kernel-impl/src/main/java/org/sakaiproject/authz/impl/DbAuthzGroupSqlMySql.java

WL-2163 Don't copy .auth/.anon roles into groups.

The problem was the .auth/.anon were getting copied into groups which means that the groups no longer function as you don't have to be a member of the group to access content protected by it.