WL-699 Fixed the access URL for citations / reading lists.
This fix was taken from the "patch" provided by Steve Smail to the corresponding
central JIRA SAK-16232. This change specifially addresses the issue - no more,
no less. There is a lot of (potentially confusing) correspondence on that
ticket, pertaining to other issues that may or may not have also fixed this
issue. However, on closer inspection, it would appear that KNL-332 does fix this
after all ;-).