WL-3119 Define a message multipart subtype
Similar to the `Content-Disposition` of an attachment. The MIME multipart
subtype allows you to specify the relationship of the message parts.

This is achieved by smuggling in the subtype in an `additionalHeader`. E.g.

    multipart-subtype: mixed

This is introduced so that you can specify a subtype `alternative` for the
presence of calendar attachments which are considered 'alternative'
renderings of the textual email message.