WL-3191 Add read access to existing roles
This code was broken during some static NPE analysis or something. It
ignores the case when role is not null. It's quite possible that a role
is already defined on a resource, but doesn't yet grant READ access, e.g.
if it has been set up using the permissions helper.