WL-3453 Drop the isSuperUser() checks.
These aren’t needed because the security service will eventually check to see if the user is an administrator.

These checks were a problem for us because sometimes we will fail a permission check for the administrator if a second factor is required and the current user doesn’t have that second factor.