contact_us_email_from_address_bug : Change from-address to the actual sender
The bug is that emails from the Contact Us tool are being shown as
from "WebLearn<weblearn-noreply@it.ox.ac.uk>" whereas they should
show the sender's email address

Here I am adding the setter interface method for the smtpFrom variable
(so that it can be called from SakaiProxy)
and a getter method for it.

The idea is that we set the basicService's smptFrom then after the email is sent we reutnr it to
its originla value.